So suu tam: ve lai giao dien moi, tab panel quan ly the (hao quang, chi so tung cap, so luong len cap)
May chu
- SoSuuTamDAO: bang radar tu them cot can_moi_cap, aura_tu_cap, yeu_cau_the, yeu_cau_cap, thu_tu; options thanh TEXT. Nap / luu / xoa, ap dung ngay cho nguoi online.
- RadarCard: cap toi da theo so luong tung cap; Card: so luong kieu int, giu cap -1 (da mo khoa).
- RadarService: goi 127 dinh dang moi (int so luong, short id chi so, cap toi da, hao quang); suuTam gop ca chong mot luot, len nhieu cap lien.
- NPoint: chi so the lay tu mau the qua addOption — dong cap 0 co ngay khi da suu tam, dong Lv.N khi dang bat va dat Lv.N (cong don).
- Hao quang the chi hien tu aura_tu_cap. Bat the khac tu tat the cu.
- Vat pham ngoai loai 33 co trong so cung gop duoc.
- Panel: tab "So suu tam" them / sua / xoa the, chon vat pham, icon, hao quang, bang chi so theo cap.

Client
- God.SoSuuTamUI: man ve moi (luoi the theo do hiem, khoa, vet sang, xem truoc quai / nhan vat dong, sao cap, thanh tien do, chi so theo cap, nut bat the), doc goi 127.
- Muc "So suu tam" o Chuc nang -> He thong; the Chuc nang cu cung mo man moi.

+++ Server/src/nro/repository/schema/GodGK.java
@@ -749,7 +749,7 @@ public class GodGK {
             dataArray = docCot(rs, "data_card");
             for (int i = 0; i < dataArray.size(); i++) {
                 JSONObject obj = (JSONObject) dataArray.get(i);
-                player.Cards.add(new Card(Short.parseShort(obj.get("id").toString()), Byte.parseByte(obj.get("amount").toString()), Byte.parseByte(obj.get("max").toString()), Byte.parseByte(obj.get("level").toString()), loadOptionCard((JSONArray) JSONValue.parse(obj.get("option").toString())), Byte.parseByte(obj.get("used").toString())));
+                player.Cards.add(new Card(Short.parseShort(obj.get("id").toString()), Integer.parseInt(obj.get("amount").toString()), Integer.parseInt(obj.get("max").toString()), Byte.parseByte(obj.get("level").toString()), loadOptionCard((JSONArray) JSONValue.parse(obj.get("option").toString())), Byte.parseByte(obj.get("used").toString())));
             }
             dataArray.clear();
